+++ pkg/sentry/platform/kvm/kvm_const_arm64.go
@@ -38,6 +38,8 @@ const (
 	_KVM_ARM64_REGS_SCTLR_EL1  = 0x603000000013c080
 	_KVM_ARM64_REGS_CPACR_EL1  = 0x603000000013c082
 	_KVM_ARM64_REGS_VBAR_EL1   = 0x603000000013c600
+	_KVM_ARM64_REGS_TIMER_CNT  = 0x603000000013df1a
+	_KVM_ARM64_REGS_CNTFRQ_EL0 = 0x603000000013df00
 )
 
 // Arm64: Architectural Feature Access Control Register EL1.
